[development/gfx_multi_engine_cleanup] cleanup of graphics initialization, I believe that's mostly done now

+++ source/1_engine/rex_renderer_core/include/rex_renderer_core/system/gpu_scorer.h
@@ -7,6 +7,6 @@
 
 namespace rex
 {
-  using GpuScorerFn = rsl::function<s32(const GpuDescription&)>;
-  s32 calc_gpu_score(const GpuDescription& gpu);
+  using GpuScorerFn = rsl::function<s64(const GpuDescription&)>;
+  s64 calc_gpu_score(const GpuDescription& gpu);
 } // namespace rex
