+++ source/1_engine/rex_renderer_core/src/system/gpu_scorer.cpp
@@ -2,7 +2,7 @@
 
 namespace rex
 {
-  s32 calc_gpu_score(const GpuDescription& gpu)
+  s64 calc_gpu_score(const GpuDescription& gpu)
   {
     return gpu.dedicated_video_memory.size_in_bytes();
   }
